feat: validacao de dados e tratamento de excecoes - parte 2

+++ src/jogadores/jogadores.service.ts
@@ -7,6 +7,7 @@ import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { CriarJogadorDto } from './dtos/criar-jogador.dto';
 import { Jogador } from './interfaces/jogador.interface';
+import { AtualizarJogadorDto } from './dtos/atualizar-jogador.dto';
 
 @Injectable()
 export class JogadoresService {
@@ -19,7 +20,7 @@ export class JogadoresService {
 
     const jogadorEncontrado = await this.jogadorModel.findOne({ email }).exec();
 
-    if (!jogadorEncontrado) {
+    if (jogadorEncontrado) {
       throw new BadRequestException(
         `Jogador com e-mail ${email} já cadastrado!`,
       );
@@ -31,16 +32,18 @@ export class JogadoresService {
 
   async atualizarJogador(
     _id: string,
-    criarJogadorDto: CriarJogadorDto,
-  ): Promise<Jogador> {
+    atualizarJogadorDto: AtualizarJogadorDto,
+  ): Promise<void> {
     const jogadorEncontrado = await this.jogadorModel.findById({ _id }).exec();
 
     if (!jogadorEncontrado) {
       throw new NotFoundException(`Jogador com id ${_id} não encontrado!`);
     }
-    return await this.jogadorModel
-      .findOneAndUpdate({ _id }, { $set: criarJogadorDto })
+    await this.jogadorModel
+      .findOneAndUpdate({ _id }, { $set: atualizarJogadorDto })
       .exec();
+
+    return Promise.resolve();
   }
 
   async consultarTodosJogadores(): Promise<Jogador[]> {
@@ -49,13 +52,20 @@ export class JogadoresService {
 
   async consultarJogadorPorId(_id: string): Promise<Jogador> {
     const jogadorEncontrado = await this.jogadorModel.findById({ _id }).exec();
+
     if (!jogadorEncontrado) {
-      throw new NotFoundException(`Jogador com email ${_id} não encontrado`);
+      throw new NotFoundException(`Jogador não encontrado`);
     }
     return jogadorEncontrado;
   }
 
   async deletarJogador(_id: string): Promise<any> {
+    const jogadorEncontrado = await this.jogadorModel.findById({ _id }).exec();
+
+    if (!jogadorEncontrado) {
+      throw new NotFoundException('Jogador não encontrado');
+    }
+
     return await this.jogadorModel.findByIdAndDelete({ _id }).exec();
   }
 }
